Added authorization for fetch config command
